pyproject.toml
[project]
name = "rag"
version = "0.1.0"
description = "RAG server with MCP integration using Google AI Studio"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "qdrant-client>=1.11.0",
    "google-generativeai>=0.8.0",
    "mcp>=1.1.0",
    "pydantic>=2.9.0",
    "pydantic-settings>=2.5.0",
    "python-dotenv>=1.0.0",
    "aiofiles>=23.2.1",
]
